+++ traffic_monitor/tools/validator-service/validator-service.go
@@ -26,7 +26,7 @@ import (
 	"fmt"
 	"github.com/apache/trafficcontrol/lib/go-tc"
 	"github.com/apache/trafficcontrol/traffic_monitor/tmcheck"
-	to "github.com/apache/trafficcontrol/traffic_ops/v3-client"
+	to "github.com/apache/trafficcontrol/traffic_ops/v4-client"
 	"io"
 	"net/http"
 	"sort"